# Support IntEnum values as arguments to JAX functions.