@jekbradbury 2019-12-11
This is the kind of change where it might be nice to see benchmark impacts—I’m guessing it’s not especially cheap to look up the MRO?