@hawkinsp 2019-12-11
Benchmarks for `xla.abstractify`:

Before:
```
----------------------------------------------------------------------------------------------- benchmark: 8 tests ----------------------------------------------------------------------------------------------
Name (time in ns)                 Min                    Max                  Mean                StdDev                Median                IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_abstractify[arg7]       174.8903 (1.0)       1,500.2227 (1.0)        183.4349 (1.0)         29.5310 (1.0)        182.1469 (1.0)       3.8546 (1.0)     1603;1712    5,451.5249 (1.0)      199641          27
test_abstractify[3]        1,130.9749 (6.47)     19,001.0760 (12.67)    1,232.8696 (6.72)       340.8573 (11.54)    1,220.0326 (6.70)     32.9455 (8.55)      150;959      811.1158 (0.15)      50315           1
test_abstractify[3.5]      1,151.9296 (6.59)     17,426.9080 (11.62)    1,248.1737 (6.80)       374.5550 (12.68)    1,233.0711 (6.77)     35.9723 (9.33)      197;805      801.1706 (0.15)      51089           1
test_abstractify[arg5]     2,368.9354 (13.55)    19,376.0497 (12.92)    2,543.2515 (13.86)      529.6191 (17.93)    2,519.9261 (13.83)    57.1599 (14.83)    286;1691      393.1975 (0.07)      67619           1
test_abstractify[arg4]     2,414.1045 (13.80)    29,324.9032 (19.55)    2,557.2832 (13.94)      573.8472 (19.43)    2,526.9110 (13.87)    51.9212 (13.47)    303;1051      391.0400 (0.07)      46228           1
test_abstractify[arg3]     3,295.9506 (18.85)    23,075.0302 (15.38)    3,563.2614 (19.43)      928.1513 (31.43)    3,450.0845 (18.94)    81.8400 (21.23)     590;657      280.6418 (0.05)      30862           1
test_abstractify[arg2]     3,310.9682 (18.93)    28,777.9840 (19.18)    3,641.9705 (19.85)    1,206.9753 (40.87)    3,466.0334 (19.03)    71.9447 (18.66)    827;1420      274.5766 (0.05)      32159           1
test_abstractify[arg6]     3,311.8995 (18.94)    21,117.0409 (14.08)    3,506.2031 (19.11)      655.4896 (22.20)    3,471.9706 (19.06)    78.1147 (20.27)     299;428      285.2088 (0.05)      38732           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

```

After:
```
------------------------------------------------------------------------------------------------ benchmark: 9 tests -----------------------------------------------------------------------------------------------                                                                                                                                                                                                                      
Name (time in ns)                     Min                    Max                  Mean              StdDev                Median                IQR            Outliers  OPS (Kops/s)            Rounds  Iterations                                                                                                                                                                                                                      
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------                                                                                                                                                                                                                      
test_abstractify[arg7]           339.7465 (1.0)       1,897.5988 (1.0)        354.8134 (1.0)       52.1613 (1.0)        351.4462 (1.0)       7.7940 (1.0)     1964;2091    2,818.3832 (1.0)      137154          20                                                                                                                                                                                                                      
test_abstractify[3]            1,344.9462 (3.96)     17,700.0184 (9.33)     1,463.5935 (4.12)     384.5724 (7.37)     1,437.9621 (4.09)     34.9246 (4.48)      90;2931      683.2498 (0.24)      39332           1                                                                                                                                                                                                                      
test_abstractify[3.5]          1,348.0894 (3.97)     17,734.0116 (9.35)     1,474.6307 (4.16)     412.3193 (7.90)     1,446.9260 (4.12)     32.9455 (4.23)     513;2511      678.1359 (0.24)      44098           1                                                                                                                                                                                                                      
test_abstractify[AnEnum.B]     1,530.9779 (4.51)     17,912.0107 (9.44)     1,681.2049 (4.74)     489.0413 (9.38)     1,657.9870 (4.72)     43.0737 (5.53)      275;625      594.8115 (0.21)      49784           1                                                                                                                                                                                                                      
test_abstractify[arg5]         2,634.9444 (7.76)     21,612.9702 (11.39)    2,813.3402 (7.93)     520.4653 (9.98)     2,779.9979 (7.91)     40.1633 (5.15)     645;3227      355.4494 (0.13)      60205           1                                                                                                                                                                                                                      
test_abstractify[arg4]         2,661.9527 (7.84)     35,081.0587 (18.49)    2,825.6567 (7.96)     548.3394 (10.51)    2,796.0632 (7.96)     39.1155 (5.02)     340;1324      353.9000 (0.13)      37039           1                                                                                                                                                                                                                      
test_abstractify[arg6]         3,585.0098 (10.55)    20,134.9612 (10.61)    3,881.5588 (10.94)    816.7873 (15.66)    3,768.0147 (10.72)    55.0644 (7.06)     925;2809      257.6285 (0.09)      36897           1                                                                                                                                                                                                                      
test_abstractify[arg2]         3,608.0601 (10.62)    20,427.0473 (10.76)    3,864.9442 (10.89)    586.9463 (11.25)    3,812.9510 (10.85)    60.0703 (7.71)     962;1702      258.7359 (0.09)      28688           1                                                                                                                                                                                                                      
test_abstractify[arg3]         3,621.9135 (10.66)    20,258.9436 (10.68)    3,843.0848 (10.83)    638.8736 (12.25)    3,779.0742 (10.75)    54.0167 (6.93)     617;1337      260.2076 (0.09)      27738           1                                                                                                                                                                                                                      
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------                                                                                                                                                                                                                      
```

i.e., there's a slowdown, but it seems to be in the 200-300ns range.

I added the benchmark to the PR.